# 6x faster IRR