On the Learn repo's IRR benchmark:

Before:
```julia-repl
julia> @benchmark irr($cfs)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  93.083 μs … 140.500 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     93.292 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   94.507 μs ±   2.504 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

  █▅▁     ▇▃  ▂▂▁▃    ▁▁▁                                      ▂
  █████▇▇▇███████████▇█████▇▇▇▇▇▇█▇▇▆▄▅▃▅▄▄▄▅▄▁▃▃▄▄▅▅▄▁▃▅▄▅▅▅▄ █
  93.1 μs       Histogram: log(frequency) by time       106 μs <

 Memory estimate: 32 bytes, allocs estimate: 1.
```

After: 
```julia-repl
 julia> @benchmark irr($cfs)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  15.541 μs …  44.250 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     15.666 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   15.773 μs ± 974.644 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

   ▅ █ █▆ ▄ ▁    ▂ ▃▁      ▁ ▂                       ▁         ▂
  ▇█▁█▁██▁█▁█▁▅▇▁█▁██▁▇▁▆▁▆█▁█▁▇▆▁▆▁▇▄▆▆▁▆▁▅▆▁▆▁▆▇▄█▁█▁█▇▁▆▁▇▅ █
  15.5 μs       Histogram: log(frequency) by time        17 μs <

 Memory estimate: 32 bytes, allocs estimate: 1.
```